migrate superfamily/list to universal profile
(merely informative check)
